+++ app/collections/rhine-river-cruise/page.tsx
@@ -22,28 +22,20 @@ const otherCollections = [
     "title": "Janet Buys a car"
   },
   {
-    "slug": "nova-scotia",
-    "title": "Nova Scotia"
-  },
-  {
     "slug": "norway-cruise",
     "title": "Norway Cruise"
-  },
-  {
-    "slug": "katie-ben-s-wedding",
-    "title": "Katie & Ben's Wedding"
-  },
-  {
-    "slug": "france-trip",
-    "title": "France trip"
   },
   {
     "slug": "blue-yoga-nyla",
     "title": "Blue Yoga Nyla"
   },
   {
-    "slug": "submit-test",
-    "title": "Submit Test"
+    "slug": "kyle",
+    "title": "Kyle"
+  },
+  {
+    "slug": "nova-scotia",
+    "title": "Nova Scotia"
   }
 ];